pwm: meson: mark the lo and hi as u16
The values in hardware are only 16-bit wide. Make this easy to see when
reading the code (instead of having to assume that this is the case by
looking at MISC_B_CLK_DIV_SHIFT or the datasheet). No functional changes
intended.

+++ drivers/pwm/pwm-meson.c
@@ -90,8 +90,8 @@ static const unsigned int mux_reg_shifts[] = {
 };
 
 struct meson_pwm_channel {
-	unsigned int hi;
-	unsigned int lo;
+	u16 hi;
+	u16 lo;
 	u8 pre_div;
 
 	struct pwm_state state;
